Struct diem_client::Client

pub struct Client { /* fields omitted */ }

Implementations

impl Client

pub fn new<T: Into<String>>(url: T) -> Self

pub fn new_with_retry<T: Into<String>>(url: T, retry: Retry) -> Self

pub fn last_known_state(&self) -> Option<State>

pub async fn wait_for_signed_transaction(
&self,
txn: &SignedTransaction,
timeout: Option<Duration>,
delay: Option<Duration>
) -> Result<Response<TransactionView>, WaitForTransactionError>

pub async fn wait_for_transaction(
&self,
address: AccountAddress,
seq: u64,
expiration_time_secs: u64,
txn_hash: HashValue,
timeout: Option<Duration>,
delay: Option<Duration>
) -> Result<Response<TransactionView>, WaitForTransactionError>

pub async fn batch(
&self,
requests: Vec<MethodRequest>
) -> Result<Vec<Result<Response<MethodResponse>>>>

pub async fn request(
&self,
request: MethodRequest
) -> Result<Response<MethodResponse>>

pub async fn submit(&self, txn: &SignedTransaction) -> Result<Response<()>>

pub async fn get_metadata_by_version(
&self,
version: u64
) -> Result<Response<MetadataView>>

pub async fn get_metadata(&self) -> Result<Response<MetadataView>>

pub async fn get_account(
&self,
address: AccountAddress
) -> Result<Response<Option<AccountView>>>

pub async fn get_account_by_version(
&self,
address: AccountAddress,
version: u64
) -> Result<Response<Option<AccountView>>>

pub async fn get_transactions(
&self,
start_seq: u64,
limit: u64,
include_events: bool
) -> Result<Response<Vec<TransactionView>>>

pub async fn get_account_transaction(
&self,
address: AccountAddress,
seq: u64,
include_events: bool
) -> Result<Response<Option<TransactionView>>>

pub async fn get_account_transactions(
&self,
address: AccountAddress,
start_seq: u64,
limit: u64,
include_events: bool
) -> Result<Response<Vec<TransactionView>>>

pub async fn get_events(
&self,
key: EventKey,
start_seq: u64,
limit: u64
) -> Result<Response<Vec<EventView>>>

pub async fn get_currencies(&self) -> Result<Response<Vec<CurrencyInfoView>>>

pub async fn get_network_status(&self) -> Result<Response<u64>>

pub async fn get_state_proof(
&self,
from_version: u64
) -> Result<Response<StateProofView>>

pub async fn get_accumulator_consistency_proof(
&self,
client_known_version: Option<u64>,
ledger_version: Option<u64>
) -> Result<Response<AccumulatorConsistencyProofView>>

pub async fn get_account_state_with_proof(
&self,
address: AccountAddress,
from_version: Option<u64>,
to_version: Option<u64>
) -> Result<Response<AccountStateWithProofView>>

pub async fn get_transactions_with_proofs(
&self,
start_version: u64,
limit: u64,
include_events: bool
) -> Result<Response<Option<TransactionsWithProofsView>>>

pub async fn get_account_transactions_with_proofs(
&self,
address: AccountAddress,
start_seq: u64,
limit: u64,
include_events: bool,
ledger_version: Option<u64>
) -> Result<Response<AccountTransactionsWithProofView>>

pub async fn get_events_with_proofs(
&self,
key: EventKey,
start_seq: u64,
limit: u64
) -> Result<Response<Vec<EventWithProofView>>>

pub async fn get_event_by_version_with_proof(
&self,
key: EventKey,
version: Option<u64>
) -> Result<Response<EventByVersionWithProofView>>

pub async fn get_deserialized_events<T: MoveStructType + DeserializeOwned>(
&self,
event_key: &EventKey,
start_seq: u64,
limit: u64
) -> Result<Response<Vec<Event<T>>>>

Return the events of type T that have been emitted to event_key since start_seq, with a max of limit results Returns an empty vector if there are no such events The type T must match the event types associated with event_key

pub async fn get_deserialized_resource<T: MoveResource>(
&self,
address: AccountAddress
) -> Result<Response<Option<T>>>

Deserialize and return the resource value of type T stored under address Returns None if there is no such value